+++ .gitignore
@@ -1,0 +1,10 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+.cache
+
+# Developer
+.project
+.pydevproject
+.settings/
+*.tar
