+++ README.md
@@ -1,6 +1,24 @@
-You'll need to update all the following per your prefs:
+katello-rapidnode
+=================
 
-* katello-rapidnode.ini (with proper values)
-* myrepofile.repo (with any additional repos you want to install)
+To run:
 
-Then execute 'katello-rapidnode.py'
+1. `git clone` this repository
+2. Update the following per your prefs:
+	* `katello-rapidnode.ini`
+	* `myrepofile.repo` (with any additional repos you want to install)
+3. Execute `katello-rapidnode.py`
+
+
+Code Contribution
+-----------------
+
+1. git clone this repo
+2. Make necessary code changes
+3. Run local tests and validate the code
+4. Make sure pylint and flake8 pass
+
+```sh
+flake8 .
+pylint *.py
+```
