+++ katello_rapidnode.ini
